♻️ GH-143 Consolidate audit log reading into audit/log_reader
- Extract iter_records/summarize/prune + path helpers and the
  append_record write API into src/dev10x/audit/log_reader.py
- Move @audit_hook decorator and wrap-record helpers to
  src/dev10x/hooks/audit_emit.py; writes now delegate to
  log_reader.append_record so the reader owns log layout
- Drop the duplicate JSONL parser in audit/__init__.py:hook_recent
  — it now calls log_reader.iter_records with explicit paths
- Update commands/hook.py to import from dev10x.audit; update all
  hook scripts (session-start/stop, skill-pre/post, ruff-format,
  audit-record) to import from dev10x.hooks.audit_emit
- Delete the old src/dev10x/hooks/audit.py module

Refs: docs/memos/004-2026-05-09-architecture-audit.md (M7 #D1/#I8, T5)
Fixes: https://github.com/Dev10x-Guru/Dev10x-Claude/issues/143
